+++ test/rates/rate-controller.js
@@ -3310,6 +3310,39 @@ function describePidControllerUpdateTests(deployFunc, getController) {
         expect(currentRate).to.equal(startingRate);
     });
 
+    it("Manually pushing a rate does not result in kickback with the next update", async function () {
+        const controller = await getController();
+
+        const period = await controller.period();
+
+        // Set input and target to the same 90% (using 8 decimals of input precision)
+        const targetRate = ethers.utils.parseUnits("0.90", 8);
+        const inputRate = ethers.utils.parseUnits("0.10", 8);
+        await controller.setTarget(GRT, targetRate);
+        await controller.setInput(GRT, inputRate);
+
+        const startingRate = ethers.utils.parseUnits("0.234", 8);
+        await controller.manuallyPushRate(GRT, startingRate, startingRate, 1);
+
+        // Advance the block time by the period
+        await timeAndMine.increaseTime(period.toNumber());
+
+        // Update the rate
+        await controller.update(ethers.utils.defaultAbiCoder.encode(["address"], [GRT]));
+
+        // Get the current rate
+        const currentRate = await controller.computeRate(GRT);
+
+        // Calculate expected rate: The proportional component should not kick
+        const error = targetRate.sub(inputRate);
+        const expectedChange = error.mul(DEFAULT_PID_CONFIG.kINumerator).div(DEFAULT_PID_CONFIG.kIDenominator);
+        const expectedRate = startingRate.add(expectedChange);
+
+        // Confirm that the new rate equals the manually pushed rate
+        // If the PID controller is not reinitialized, the rate will be the same as rate before the manually pushed rate (0)
+        expect(currentRate).to.equal(expectedRate);
+    });
+
     it("It works with a transformer", async function () {
         const deployment1 = await deployFunc();
         const deployment2 = await deployFunc();
@@ -4128,6 +4161,7 @@ function createDescribeCanUpdateTests(isPidController) {
 function describeTests(
     contractName,
     deployFunc,
+    setDefaultConfigFunc,
     describeComputeRateTests,
     describeNeedsUpdateTests,
     describeUpdateTests,
@@ -4250,12 +4284,16 @@ function describeTests(
             await expect(controller.connect(signer).setUpdatesPaused(GRT, true)).to.not.be.reverted;
         });
 
-        it("Should revert if the token is missing a config", async function () {
-            await expect(controller.setUpdatesPaused(USDC, true)).to.be.revertedWith("MissingConfig").withArgs(USDC);
+        it("Should not revert if the token is missing a config", async function () {
+            await expect(controller.setUpdatesPaused(USDC, true))
+                .to.emit(controller, "PauseStatusChanged")
+                .withArgs(USDC, true);
 
-            // Sanity check that we can successfully call the function if we have the config
-            await controller.setConfig(USDC, DEFAULT_CONFIG);
-            await expect(controller.setUpdatesPaused(USDC, true)).to.not.be.reverted;
+            // Sanity check that the changes were made
+            expect(await controller.areUpdatesPaused(USDC)).to.equal(true);
+
+            // Sanity check that USDC is missing a config
+            await expect(controller.getConfig(USDC)).to.be.revertedWith("MissingConfig").withArgs(USDC);
         });
 
         it("Should emit an event when the updates are paused", async function () {
@@ -5659,7 +5697,7 @@ function describeTests(
             await controller.grantRole(UPDATE_PAUSE_ADMIN_ROLE, signer.address);
 
             // Set config for GRT
-            await controller.setConfig(GRT, DEFAULT_CONFIG);
+            await setDefaultConfigFunc(controller, GRT);
         });
 
         it("Should revert if the caller does not have the ADMIN role", async function () {
@@ -5690,11 +5728,11 @@ function describeTests(
             const rate = ethers.utils.parseUnits("0.1234", 18);
 
             await expect(controller.manuallyPushRate(USDC, rate, rate, 1))
-                .to.be.revertedWith("MissingConfig")
+                .to.be.revertedWith(/MissingPidConfig|MissingConfig/)
                 .withArgs(USDC);
 
             // Sanity check that it works if we set a config
-            await controller.setConfig(USDC, DEFAULT_CONFIG);
+            await setDefaultConfigFunc(controller, USDC);
             await expect(controller.manuallyPushRate(USDC, rate, rate, 1)).to.not.be.reverted;
         });
 
@@ -5811,9 +5849,19 @@ async function initializePidController(controller) {
     await controller.setPidConfig(GRT, DEFAULT_PID_CONFIG);
 }
 
+async function setDefaultStandardConfig(controller, token) {
+    await controller.setConfig(token, DEFAULT_CONFIG);
+}
+
+async function setDefaultPidConfig(controller, token) {
+    await controller.setConfig(token, DEFAULT_CONFIG);
+    await controller.setPidConfig(token, DEFAULT_PID_CONFIG);
+}
+
 describeTests(
     "RateController",
     deployStandardController,
+    setDefaultStandardConfig,
     describeStandardControllerComputeRateTests,
     createDescribeStandardControllerNeedsUpdateTests(false, undefined, undefined),
     createDescribeStandardControllerUpdateTests(undefined, true, undefined),
@@ -5822,6 +5870,7 @@ describeTests(
 describeTests(
     "PidController",
     deployPidController,
+    setDefaultPidConfig,
     describePidControllerComputeRateTests,
     createDescribeStandardControllerNeedsUpdateTests(
         true,